Modify behaviour of CupsCtrl::lpdqIsValid test method
Previously, the test for valid characters in an LPD queue name was
very restrictive, allowing only CHARS_NUMERIC_ALPHA. RFC 1179 makes no
such restrictions.

This change adds '_' and '-' to the allowed characters. Other
characters can cause issues in the web interface, so I'm keeping it
restrictive, but this now matches the character which we allow in
CUPS/IPP queues.

This change also adds some additional code documentation in the header.